+++ src/main/java/pl/plajer/murdermystery/arena/ArenaEvents.java
@@ -1,6 +1,6 @@
 /*
  * MurderMystery - Find the murderer, kill him and survive!
- * Copyright (C) 2019  Plajer's Lair - maintained by Tigerpanzer_02, Plajer and contributors
+ * Copyright (C) 2020  Plugily Projects - maintained by Tigerpanzer_02, 2Wild4You and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -396,6 +396,7 @@ public class ArenaEvents implements Listener {
       player.setGameMode(GameMode.SURVIVAL);
       player.removePotionEffect(PotionEffectType.NIGHT_VISION);
       user.setStat(StatsStorage.StatisticType.LOCAL_GOLD, 0);
+      plugin.getRewardsHandler().performReward(player, Reward.RewardType.DEATH);
     }
   }